Variable Property Columns v1
One can link a session to a selected project column and determine which column holds the date time.

Setup is now split into:
Config, which happens in VS Code settings and are thing that rarely need to change
Settings, which could (though often not) change more regularly and are usually dependent on each other, editable through a CodeTrack Panel

Structure has changed slightly for tidiness and Logging is now more structured, which each file having a combo of identifiers to locate where a message comes from.